+++ miden-lib/src/transaction/inputs.rs
@@ -131,7 +131,12 @@ fn build_advice_stack(
     inputs.extend_stack(header.nullifier_root());
     inputs.extend_stack(header.batch_root());
     inputs.extend_stack(header.proof_hash());
-    inputs.extend_stack([header.block_num().into(), header.version(), header.timestamp(), ZERO]);
+    inputs.extend_stack([
+        header.block_num().into(),
+        header.version().into(),
+        header.timestamp().into(),
+        ZERO,
+    ]);
     inputs.extend_stack([ZERO; 4]);
     inputs.extend_stack(header.note_root());
 
@@ -217,7 +222,7 @@ fn add_account_to_advice_inputs(
     )]);
 
     // If there are storage maps, we populate the merkle store and advice map
-    if !(account.storage().maps().is_empty()) {
+    if !account.storage().maps().is_empty() {
         for map in account.storage().maps() {
             // extend the merkle store and map with the storage maps
             inputs.extend_merkle_store(map.inner_nodes());