+++ miden-tx/tests/integration/main.rs
@@ -140,6 +140,7 @@ pub fn prove_and_verify_transaction(
 ) -> Result<(), TransactionVerifierError> {
     let executed_transaction_id = executed_transaction.id();
     // Prove the transaction
+
     let proof_options = ProvingOptions::default();
     let prover = TransactionProver::new(proof_options);
     let proven_transaction = prover.prove_transaction(executed_transaction).unwrap();
@@ -157,18 +158,25 @@ pub fn prove_and_verify_transaction(
 }
 
 #[cfg(test)]
-pub fn get_new_key_pair_with_advice_map() -> (Word, Vec<Felt>) {
+pub fn get_new_pk_and_authenticator(
+) -> (Word, std::rc::Rc<miden_tx::host::BasicAuthenticator<rand::rngs::StdRng>>) {
+    use std::rc::Rc;
+
+    use miden_tx::host::BasicAuthenticator;
+    use rand::rngs::StdRng;
+
     let seed = [0_u8; 32];
     let mut rng = ChaCha20Rng::from_seed(seed);
 
     let sec_key = SecretKey::with_rng(&mut rng);
     let pub_key: Word = sec_key.public_key().into();
-    let mut pk_sk_bytes = sec_key.to_bytes();
-    pk_sk_bytes.append(&mut pub_key.to_bytes());
-    let pk_sk_felts: Vec<Felt> =
-        pk_sk_bytes.iter().map(|a| Felt::new(*a as u64)).collect::<Vec<Felt>>();
 
-    (pub_key, pk_sk_felts)
+    let authenticator = BasicAuthenticator::<StdRng>::new(&[(
+        pub_key,
+        miden_tx::host::AuthSecretKey::RpoFalcon512(sec_key),
+    )]);
+
+    (pub_key, Rc::new(authenticator))
 }
 
 #[cfg(test)]