+++ objects/src/lib.rs
@@ -22,8 +22,9 @@ mod errors;
 pub use block::BlockHeader;
 pub use constants::*;
 pub use errors::{
-    AccountDeltaError, AccountError, AssetError, AssetVaultError, ChainMmrError, NoteError,
-    ProvenTransactionError, TransactionInputError, TransactionOutputError, TransactionScriptError,
+    AccountDeltaError, AccountError, AssetError, AssetVaultError, BlockError, ChainMmrError,
+    NoteError, ProvenTransactionError, TransactionInputError, TransactionOutputError,
+    TransactionScriptError,
 };
 pub use miden_crypto::hash::rpo::{Rpo256 as Hasher, RpoDigest as Digest};
 pub use vm_core::{Felt, FieldElement, StarkField, Word, EMPTY_WORD, ONE, WORD_SIZE, ZERO};