+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * [BREAKING] Removed the transaction script root output from the transaction kernel (#608).
 * [BREAKING] Refactored account update details, moved `Block` to `miden-objects` (#618, #621).
-* [BREAKING] Made initial hash optional (`None` for new accounts) (#529)
+* [BREAKING] Changed type of `version` and `timestamp` fields to `u32`, moved `version` to the beginning of block header
+  (#639).
 * Introduce the `miden-bench-tx` crate used for transactions benchmarking (#577).
 
 ## 0.2.3 (2024-04-26) - `miden-tx` crate only